Fix duplicate log messages
Because the syslog logger was a child of the root logger, it inherited
the stream_handler from the root logger. This lead to duplicate log
messages when world writable files were detected. A more granular
approach to configuring loggers has been adopted to improve formatting
and remove the duplicate log messages.