Prefix watch directories with snap hostfs path
Because the snap package uses the system-files interface, all system
files are accessible at the path "/var/lib/snap/hostfs". Since this is
cumbersome for a user to remember and type whenever they use owwatcher.
Detect whether or not this application is running as a snap package and
prefix the requisite path so the user doesn't have to.