main.py: Change the scope of logger_configurer so loggers don't get cleaned up
The python logging module creates singleton loggers. Each unit test must
clean up the loggers after it runs so that the next unit test doesn't
inherit a singleton logger that's already configured with handlers. This
cleanup is being handled in OWWatcherLoggerConfigurer.__del__().

When main configures loggers, it must ensure that the
OWWatcherLoggerConfigurer instance does not go out of scope, otherwise
it's __del__() will be called and wipe away the logger configuration.